implemented create-source-stream-mapping destinationConfig for dynamodb streams; fixed lambda-input from dynamodb stream;  added test cases; refactored tests
