+++ tests/integration/awslambda/test_lambda_integration.py
@@ -13,18 +13,24 @@ from localstack.services.awslambda.lambda_api import (
     BATCH_SIZE_RANGES,
     INVALID_PARAMETER_VALUE_EXCEPTION,
 )
-from localstack.services.awslambda.lambda_utils import LAMBDA_RUNTIME_PYTHON36
+from localstack.services.awslambda.lambda_utils import (
+    LAMBDA_RUNTIME_PYTHON36,
+    LAMBDA_RUNTIME_PYTHON37,
+)
 from localstack.utils import testutil
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import retry, safe_requests, short_uid
+from localstack.utils.strings import to_bytes
 from localstack.utils.sync import poll_condition
 from localstack.utils.testutil import check_expected_lambda_log_events_length, get_lambda_log_events
 
+from .functions import lambda_integration
 from .test_lambda import (
     TEST_LAMBDA_FUNCTION_PREFIX,
     TEST_LAMBDA_LIBS,
     TEST_LAMBDA_PYTHON,
     TEST_LAMBDA_PYTHON_ECHO,
+    TEST_LAMBDA_PYTHON_UNHANDLED_ERROR,
     is_old_provider,
 )
 
@@ -33,6 +39,44 @@ TEST_SNS_TOPIC_NAME = "sns-topic-1"
 
 THIS_FOLDER = os.path.dirname(os.path.realpath(__file__))
 TEST_LAMBDA_PARALLEL_FILE = os.path.join(THIS_FOLDER, "functions", "lambda_parallel.py")
+
+
+lambda_role = {
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Effect": "Allow",
+            "Principal": {"Service": "lambda.amazonaws.com"},
+            "Action": "sts:AssumeRole",
+        }
+    ],
+}
+s3_lambda_permission = {
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Effect": "Allow",
+            "Action": [
+                "sqs:*",
+                "dynamodb:DescribeStream",
+                "dynamodb:GetRecords",
+                "dynamodb:GetShardIterator",
+                "dynamodb:ListStreams",
+                "kinesis:DescribeStream",
+                "kinesis:DescribeStreamSummary",
+                "kinesis:GetRecords",
+                "kinesis:GetShardIterator",
+                "kinesis:ListShards",
+                "kinesis:ListStreams",
+                "kinesis:SubscribeToShard",
+                "logs:CreateLogGroup",
+                "logs:CreateLogStream",
+                "logs:PutLogEvents",
+            ],
+            "Resource": ["*"],
+        }
+    ],
+}
 
 
 class TestLambdaEventSourceMappings:
@@ -395,6 +439,308 @@ class TestLambdaEventSourceMappings:
 
         assert "Subject" in notification
         assert subject == notification["Subject"]
+
+    def test_event_source_mapping_with_destination_config_kinesis_stream(
+        self,
+        lambda_client,
+        create_lambda_function,
+        sqs_client,
+        sqs_queue_arn,
+        sqs_create_queue,
+        iam_create_role_with_policy,
+        kinesis_client,
+    ):
+        try:
+            function_name = f"lambda_func-{short_uid()}"
+            role = f"test-lambda-role-{short_uid()}"
+            policy_name = f"test-lambda-policy-{short_uid()}"
+            role_arn = iam_create_role_with_policy(
+                RoleName=role,
+                PolicyName=policy_name,
+                RoleDefinition=lambda_role,
+                PolicyDefinition=s3_lambda_permission,
+            )
+
+            response = create_lambda_function(
+                handler_file=TEST_LAMBDA_PYTHON,
+                func_name=function_name,
+                runtime=LAMBDA_RUNTIME_PYTHON37,
+                role=role_arn,
+                libs=TEST_LAMBDA_LIBS,
+            )
+
+            response["CreateFunctionResponse"]["FunctionArn"]
+
+            # test destination-config for create_event_source_mapping
+            kinesis_name = f"test-kinesis-{short_uid()}"
+            kinesis_client.create_stream(StreamName=kinesis_name, ShardCount=1)
+
+            result = kinesis_client.describe_stream(StreamName=kinesis_name)["StreamDescription"]
+            kinesis_arn = result["StreamARN"]
+
+            # wait for stream-status "ACTIVE"
+            status = result["StreamStatus"]
+            if status != "ACTIVE":
+
+                def check_stream_active():
+                    state = kinesis_client.describe_stream(StreamName=kinesis_name)[
+                        "StreamDescription"
+                    ]["StreamStatus"]
+                    if state != "ACTIVE":
+                        raise Exception(f"StreamStatus is {state}")
+
+                retry(check_stream_active, retries=6, sleep=1.0, sleep_before=2.0)
+
+            queue_event_source_mapping = sqs_create_queue()
+            queue_failure_event_source_mapping_arn = sqs_queue_arn(queue_event_source_mapping)
+            destination_config = {
+                "OnFailure": {"Destination": queue_failure_event_source_mapping_arn}
+            }
+
+            result = lambda_client.create_event_source_mapping(
+                FunctionName=function_name,
+                BatchSize=1,
+                StartingPosition="TRIM_HORIZON",
+                EventSourceArn=kinesis_arn,
+                MaximumBatchingWindowInSeconds=1,
+                MaximumRetryAttempts=1,
+                DestinationConfig=destination_config,
+            )
+
+            event_source_mapping_uuid = result["UUID"]
+            event_source_mapping_state = result["State"]
+            if event_source_mapping_state != "Enabled":
+
+                def check_mapping_state():
+                    state = lambda_client.get_event_source_mapping(UUID=event_source_mapping_uuid)[
+                        "State"
+                    ]
+                    if state != "Enabled":
+                        raise Exception(f"State is {state}")
+
+                retry(check_mapping_state, retries=6, sleep_before=2.0, sleep=1.0)
+
+            message = {
+                "input": "hello",
+                "value": "world",
+                lambda_integration.MSG_BODY_RAISE_ERROR_FLAG: 1,
+            }
+            kinesis_client.put_record(
+                StreamName=kinesis_name, Data=to_bytes(json.dumps(message)), PartitionKey="custom"
+            )
+
+            def verify_failure_received():
+                result = sqs_client.receive_message(QueueUrl=queue_event_source_mapping)
+                msg = result["Messages"][0]
+                body = json.loads(msg["Body"])
+                assert body["requestContext"]["condition"] == "RetryAttemptsExhausted"
+                assert body["KinesisBatchInfo"]["batchSize"] == 1
+                assert body["KinesisBatchInfo"]["streamArn"] == kinesis_arn
+
+            retry(verify_failure_received, retries=50, sleep=5, sleep_before=5)
+
+        finally:
+            kinesis_client.delete_stream(StreamName=kinesis_name, EnforceConsumerDeletion=True)
+            lambda_client.delete_event_source_mapping(UUID=event_source_mapping_uuid)
+
+    def test_event_source_mapping_with_destination_config_dynamodb(
+        self,
+        lambda_client,
+        create_lambda_function,
+        sqs_client,
+        sqs_queue_arn,
+        sqs_create_queue,
+        iam_create_role_with_policy,
+        dynamodb_client,
+        dynamodb_create_table,
+    ):
+        try:
+            function_name = f"lambda_func-{short_uid()}"
+            role = f"test-lambda-role-{short_uid()}"
+            policy_name = f"test-lambda-policy-{short_uid()}"
+            role_arn = iam_create_role_with_policy(
+                RoleName=role,
+                PolicyName=policy_name,
+                RoleDefinition=lambda_role,
+                PolicyDefinition=s3_lambda_permission,
+            )
+
+            create_lambda_function(
+                handler_file=TEST_LAMBDA_PYTHON_UNHANDLED_ERROR,
+                # handler_file=TEST_LAMBDA_PYTHON_ECHO,
+                func_name=function_name,
+                runtime=LAMBDA_RUNTIME_PYTHON37,
+                role=role_arn,
+                libs=TEST_LAMBDA_LIBS,
+            )
+
+            # test destination-config for create_event_source_mapping
+            table_name = f"test-table-{short_uid()}"
+
+            partition_key = "my_partition_key"
+            result = dynamodb_create_table(table_name=table_name, partition_key=partition_key)
+
+            status = result["TableDescription"]["TableStatus"]
+            # wait for status "ACTIVE"
+            if status != "ACTIVE":
+
+                def check_table_active():
+                    state = dynamodb_client.describe_table(TableName=table_name)["Table"][
+                        "TableStatus"
+                    ]
+                    if state != "ACTIVE":
+                        raise Exception(f"TableStatus is {state}")
+
+                retry(check_table_active, retries=6, sleep=1.0, sleep_before=2.0)
+
+            # activate stream
+            result = dynamodb_client.update_table(
+                TableName=table_name,
+                StreamSpecification={"StreamEnabled": True, "StreamViewType": "NEW_IMAGE"},
+            )
+            stream_arn = result["TableDescription"]["LatestStreamArn"]
+            queue_event_source_mapping = sqs_create_queue()
+            queue_failure_event_source_mapping_arn = sqs_queue_arn(queue_event_source_mapping)
+            destination_config = {
+                "OnFailure": {"Destination": queue_failure_event_source_mapping_arn}
+            }
+
+            result = lambda_client.create_event_source_mapping(
+                FunctionName=function_name,
+                BatchSize=1,
+                StartingPosition="TRIM_HORIZON",
+                EventSourceArn=stream_arn,
+                MaximumBatchingWindowInSeconds=1,
+                MaximumRetryAttempts=1,
+                DestinationConfig=destination_config,
+            )
+
+            event_source_mapping_uuid = result["UUID"]
+            event_source_mapping_state = result["State"]
+            if event_source_mapping_state != "Enabled":
+
+                def check_mapping_state():
+                    state = lambda_client.get_event_source_mapping(UUID=event_source_mapping_uuid)[
+                        "State"
+                    ]
+                    if state != "Enabled":
+                        raise Exception(f"State is {state}")
+
+                retry(check_mapping_state, retries=6, sleep_before=2.0, sleep=1.0)
+
+            insert = {partition_key: {"S": "hello world"}}
+
+            dynamodb_client.put_item(TableName=table_name, Item=insert)
+
+            def verify_failure_received():
+                res = sqs_client.receive_message(QueueUrl=queue_event_source_mapping)
+                msg = res["Messages"][0]
+                body = json.loads(msg["Body"])
+                assert body["requestContext"]["condition"] == "RetryAttemptsExhausted"
+                assert body["DDBStreamBatchInfo"]["batchSize"] == 1
+                assert body["DDBStreamBatchInfo"]["streamArn"] in stream_arn
+
+            retry(verify_failure_received, retries=50, sleep=5, sleep_before=5)
+
+        finally:
+            lambda_client.delete_event_source_mapping(UUID=event_source_mapping_uuid)
+
+    def test_event_source_mapping_dynamodb(
+        self,
+        lambda_client,
+        create_lambda_function,
+        iam_create_role_with_policy,
+        dynamodb_client,
+        dynamodb_create_table,
+        logs_client,
+        check_lambda_logs,
+    ):
+        try:
+            function_name = f"lambda_func-{short_uid()}"
+            role = f"test-lambda-role-{short_uid()}"
+            policy_name = f"test-lambda-policy-{short_uid()}"
+            role_arn = iam_create_role_with_policy(
+                RoleName=role,
+                PolicyName=policy_name,
+                RoleDefinition=lambda_role,
+                PolicyDefinition=s3_lambda_permission,
+            )
+
+            create_lambda_function(
+                handler_file=TEST_LAMBDA_PYTHON_ECHO,
+                func_name=function_name,
+                runtime=LAMBDA_RUNTIME_PYTHON37,
+                role=role_arn,
+                libs=TEST_LAMBDA_LIBS,
+            )
+
+            # test destination-config for create_event_source_mapping
+            table_name = f"test-table-{short_uid()}"
+
+            partition_key = "my_partition_key"
+            result = dynamodb_create_table(table_name=table_name, partition_key=partition_key)
+
+            status = result["TableDescription"]["TableStatus"]
+            # wait for status "ACTIVE"
+            if status != "ACTIVE":
+
+                def check_table_active():
+                    state = dynamodb_client.describe_table(TableName=table_name)["Table"][
+                        "TableStatus"
+                    ]
+                    if state != "ACTIVE":
+                        raise Exception(f"TableStatus is {state}")
+
+                retry(check_table_active, retries=6, sleep=1.0, sleep_before=2.0)
+
+            # activate stream
+            result = dynamodb_client.update_table(
+                TableName=table_name,
+                StreamSpecification={"StreamEnabled": True, "StreamViewType": "NEW_IMAGE"},
+            )
+            stream_arn = result["TableDescription"]["LatestStreamArn"]
+
+            result = lambda_client.create_event_source_mapping(
+                FunctionName=function_name,
+                BatchSize=1,
+                StartingPosition="TRIM_HORIZON",
+                EventSourceArn=stream_arn,
+                MaximumBatchingWindowInSeconds=1,
+                MaximumRetryAttempts=1,
+            )
+
+            event_source_mapping_uuid = result["UUID"]
+            event_source_mapping_state = result["State"]
+            if event_source_mapping_state != "Enabled":
+
+                def check_mapping_state():
+                    state = lambda_client.get_event_source_mapping(UUID=event_source_mapping_uuid)[
+                        "State"
+                    ]
+                    if state != "Enabled":
+                        raise Exception(f"State is {state}")
+
+                retry(check_mapping_state, retries=6, sleep_before=2.0, sleep=1.0)
+
+            insert = {partition_key: {"S": "hello world"}}
+
+            dynamodb_client.put_item(TableName=table_name, Item=insert)
+
+            def check_logs():
+                # assert that logs are present
+                expected = [
+                    r'.*"Records":.*',
+                    r'.*"dynamodb": {(.*)}.*',
+                    r'.*"eventSource": ("aws:dynamodb").*',
+                    r'.*"eventName": ("INSERT").*',
+                    r'.*"Keys": {("my_partition_key": {"S": "hello world"})}.*',
+                ]
+                check_lambda_logs(function_name, expected_lines=expected)
+
+            retry(check_logs, retries=50, sleep=2)
+
+        finally:
+            lambda_client.delete_event_source_mapping(UUID=event_source_mapping_uuid)
 
 
 class TestLambdaHttpInvocation:
